+++ src/quantum/azext_quantum/_params.py
@@ -23,12 +23,14 @@ def load_arguments(self, _):
     skip_role_assignment_type = CLIArgumentType(help='Skip the role assignment step for the quantum workspace in the storage account.')
     provider_id_type = CLIArgumentType(options_list=['--provider-id', '-p'], help='Identifier of an Azure Quantum provider.')
     sku_type = CLIArgumentType(options_list=['--sku', '-k'], help='Identify a plan or SKU offered by an Azure Quantum provider.')
+    provider_sku_list_type = CLIArgumentType(options_list=['--provider-sku-list', '-r'], help='Comma separated list of Provider/SKU pairs.')
 
     with self.argument_context('quantum workspace') as c:
         c.argument('workspace_name', workspace_name_type)
         c.argument('storage_account', storage_account_name_type)
         c.argument('tag', tag_type)
         c.argument('skip_role_assignment', skip_role_assignment_type)
+        c.argument('provider_sku_list', provider_sku_list_type)
 
     with self.argument_context('quantum target') as c:
         c.argument('workspace_name', workspace_name_type)