+++ src/aks-preview/azext_aks_preview/_consts.py
@@ -16,6 +16,7 @@ CONST_SPOT_EVICTION_POLICY_DEALLOCATE = "Deallocate"
 CONST_RAPID_UPGRADE_CHANNEL = "rapid"
 CONST_STABLE_UPGRADE_CHANNEL = "stable"
 CONST_PATCH_UPGRADE_CHANNEL = "patch"
+CONST_NODE_IMAGE_UPGRADE_CHANNEL = "node-image"
 CONST_NONE_UPGRADE_CHANNEL = "none"
 
 CONST_HTTP_APPLICATION_ROUTING_ADDON_NAME = "httpApplicationRouting"
@@ -59,6 +60,10 @@ CONST_ACC_SGX_QUOTE_HELPER_ENABLED = "ACCSGXQuoteHelperEnabled"
 CONST_PRIVATE_DNS_ZONE_SYSTEM = "system"
 CONST_PRIVATE_DNS_ZONE_NONE = "none"
 
+# Azure Keyvault Secrets Provider configuration keys
+CONST_AZURE_KEYVAULT_SECRETS_PROVIDER_ADDON_NAME = "azureKeyvaultSecretsProvider"
+CONST_SECRET_ROTATION_ENABLED = "enableSecretRotation"
+
 ADDONS = {
     'http_application_routing': CONST_HTTP_APPLICATION_ROUTING_ADDON_NAME,
     'monitoring': CONST_MONITORING_ADDON_NAME,
@@ -68,5 +73,6 @@ ADDONS = {
     'ingress-appgw': CONST_INGRESS_APPGW_ADDON_NAME,
     'open-service-mesh': CONST_OPEN_SERVICE_MESH_ADDON_NAME,
     "confcom": CONST_CONFCOM_ADDON_NAME,
-    'gitops': 'gitops'
+    'gitops': 'gitops',
+    'azure-keyvault-secrets-provider': CONST_AZURE_KEYVAULT_SECRETS_PROVIDER_ADDON_NAME
 }