+++ src/spring-cloud/HISTORY.md
@@ -2,6 +2,7 @@ Release History
 ===============
 2.2.1
 -----
+* Fix exception in app service binding
 * Support End-to-end TLS.
 
 2.2.0