+++ src/spring-cloud/HISTORY.md
@@ -1,8 +1,18 @@
 Release History
 ===============
+2.2.1
+-----
+* Fix exception in app service binding
+
+2.2.0
+-----
+* Add bring your own route tables support
+
 2.1.2
 -----
 * Add optional '--deployment' to 'az spring-cloud app logs' command
+* Add a parameter '--assign-endpoint' into 'az spring-cloud app create' and 'az spring-cloud app update'
+* Deprecate the parameter '--is-public' in 'az spring-cloud app create' and 'az spring-cloud app update'
 
 2.1.1
 -----